+++ kamu-core/src/infra/engine/mod.rs
@@ -7,8 +7,8 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-mod engine_provisioner_impl;
-pub use engine_provisioner_impl::*;
+mod engine_provisioner_local;
+pub use engine_provisioner_local::*;
 
 mod engine_odf;
 mod engine_spark;